Create an abstract class Publisher to simplify publishers
