+++ src/Illuminate/Foundation/Console/AssetPublishCommand.php
@@ -1,12 +1,15 @@
 <?php namespace Illuminate\Foundation\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Console\ConfirmableTrait;
 use Symfony\Component\Finder\Finder;
 use Illuminate\Foundation\AssetPublisher;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 
 class AssetPublishCommand extends Command {
+
+	use ConfirmableTrait;
 
 	/**
 	 * The console command name.
@@ -63,6 +66,13 @@ class AssetPublishCommand extends Command {
 	 */
 	protected function publishAssets($package)
 	{
+		$proceed = $this->confirmToProceed('Assets Already Published!', function() use ($package)
+		{
+			return $this->assets->alreadyPublished($package);
+		});
+
+		if ( ! $proceed) return;
+
 		if ( ! is_null($path = $this->getPath()))
 		{
 			$this->assets->publish($package, $path);
